requirements.txt
# Main project requirements
boto3>=1.26.0
sagemaker>=2.150.0
numpy>=1.23.0
pandas>=1.5.0
pyyaml>=6.0
pytest>=7.0.0
flake8>=5.0.0
black>=22.0.0